See if this works
- See if (uint) will make sure the new HomesEstate is added to the database.

NOTE: If this works then this means when we add a new system estate we need to add (uint) to increment the EstateID number in the database for the Primary Key

+++ Universe/Services/DataService/Connectors/Local/LocalEstateConnector.cs
@@ -143,7 +143,7 @@ namespace Universe.Services.DataService
                 es.EstateID = Constants.MainlandEstateID;
             } else if ( (es.EstateOwner == (UUID) Constants.GovernorUUID))         // Universe Homes Estate?
             {
-                es.EstateID = Constants.HomesEstateID;
+                es.EstateID = (uint) Constants.HomesEstateID;
             } else if ( (es.EstateOwner == (UUID) Constants.RealEstateOwnerUUID) )  // System Estate?
             {
                 es.EstateID = (uint) Constants.SystemEstateID;
